Move function to symbolic.py
